pyproject.toml
[tool.poetry]
name = "django-polls"
version = "0.1.0"
description = ""
authors = ["MikeIV2007 <130536460+MikeIV2007@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
httpie = "^3.2.3"
django-rest-swagger = "^2.2.0"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
